Reconstructed user settings components
- General user info decoupled into image and info setup
- Decorated UI with vuetify implementation
- Replaced user settings navbar with sidebar
- Refactored user privacy form with vuetify
